Update patch set 3
Patch Set 3:

I also wanted to mention that the Screen Recorder seems to have issues with devices that have legacy camera support enabled (media.stagefright.legacyencoder=true).  The mp4 file that is created is only a few kilobytes and unplayable.  This issue is also with the previous version of the Screen Recorder (Screencast).

http://pastebin.com/dRgqfRxj

Patch-set: 3