Update patch set 3
Patch Set 3:

About your screen recorder issue, I cannot reproduce it, I built this app on a device with hal1 hack (and so it has legacy encoder), but playback is working. Please check if the cm screen recorder app is working for you. If it doesn't, check your device media configs

Patch-set: 3