Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This will break non-android studio builds since we won't have the BuildConfig class when building in the Lineage env.
> Why do you want to strip all the package name references? (I'm talking about this and its parent commit)

We are already using precompiled BuildConfig class for aosp env (you can check other files). This is done so external projects can change the package name with minimal modifications (aapt flag for aosp and application Id for gradle). I've tested and this builds for lineage and studio with no issues.

Patch-set: 1